+++ src/CAEStreamingDecoder.hpp
@@ -35,7 +35,7 @@ public:
 	void operator delete(void* mem);
 
 	virtual bool Initialise() = 0;
-	virtual unsigned long FillBuffer(void *dest, size_t size) = 0;
+	virtual size_t FillBuffer(void *dest, size_t size) = 0;
 	virtual long GetStreamLengthMs() = 0;
 	virtual long GetStreamPlayTimeMs() = 0;
 	virtual void SetCursor(unsigned long pos) = 0;
